+++ controls-table/src/jvmMain/kotlin/TagTableDevice.kt
@@ -1,4 +1,4 @@
-package space.kscience.controls.dataplatform
+package space.kscience.controls.tagtable
 
 import kotlinx.coroutines.flow.Flow
 import space.kscience.controls.api.*