+++ controls-table/src/jvmMain/kotlin/storage/Replay.kt
@@ -1,4 +1,4 @@
-package space.kscience.controls.dataplatform.storage
+package space.kscience.controls.tagtable.storage
 
 import kotlin.time.Instant