+++ controls-table/src/jvmMain/kotlin/timeseries/TimeSeriesRows.kt
@@ -1,9 +1,9 @@
-package space.kscience.controls.dataplatform.timeseries
+package space.kscience.controls.tagtable.timeseries
 
 import kotlinx.coroutines.flow.*
 import space.kscience.attributes.SafeType
 import space.kscience.attributes.safeTypeOf
-import space.kscience.controls.dataplatform.TagTable.Companion.timeColumnHeader
+import space.kscience.controls.tagtable.TagTable.Companion.timeColumnHeader
 import space.kscience.controls.time.ValueWithTime
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.meta.MetaConverter