+++ demo/data-platform-demo/src/main/kotlin/setupPlatformTestStand.kt
@@ -1,9 +1,8 @@
 package space.kscience.controls.demo
 
 import space.kscience.controls.constructor.DeviceConfiguration
-import space.kscience.controls.dataplatform.*
 import space.kscience.controls.manager.DeviceManager
-import space.kscience.controls.storage.PlainRowsEnvelopeConverter
+import space.kscience.controls.tagtable.*
 import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.names.parseAsName
 import space.kscience.dataforge.names.plus
@@ -91,7 +90,8 @@ internal fun DeviceManager.setupPlatformTestStand(
             readInterval = 1.seconds,
             maxDuration = 30.seconds,
             compression = RowsCompression(skipUnchangedRows = true, skipUnchangedValues = true, numericDelta = 0.05),
-            rowsConverter = PlainRowsEnvelopeConverter.meta
+            separateMeta = false
+//            rowsConverterType = PlainRowsEnvelopeConverter.meta.envelopeType
         )
     )